requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

pytest>=8.0.0
httpx>=0.27.0
ruff>=0.4.0
pyyaml>=6.0.1
langchain>=0.1.0
langchain-community>=0.0.10
chromadb==0.5.23

# Phase 3 Dependencies (Postgres & Celery)
psycopg2-binary==2.9.10
sqlalchemy>=2.0.0
pgvector>=0.2.5
langchain-postgres>=0.0.12
celery>=5.4.0
redis>=5.2.00
requests>=2.31.0
sse-starlette>=2.0.0

pypdf
